Fix #17 and and some debug output (#19)
* Just comitting the debug stuff that I added.

* Forgot to push these files up.

* Fixes the crash I was seeing.

* Screwed up the write_log line here so it wasn't giving any useful output lol

* Updating specs, adding some more debug output for my testing on WSL.

* Committing a change to registry_test to easily reproduce #17.

* Committing some extra stuff...probably isn't 100% necessary, but meh, a ton more warnings being added as well xD We're getting there though lol

* This should fix #17

+++ include/registry.h
@@ -16,7 +16,7 @@ typedef struct {
      signed long int* extregisters;
 
      // Methods
-     int (*setRegister)(void*, char, int);
+     int (*setRegister)(void*, char, unsigned long int);
      int (*performOperation)(void*, char, signed int*, int);
      int (*getRegisterValue)(void*, char, signed int*);
      signed int* (*translateRegister)(void*, char);
@@ -33,7 +33,7 @@ void Registry_construct(void* r, signed int* bp, signed int* sp, signed int* ip,
 void Registry_destruct(void* r);
 
 // Registry object methods
-int Registry_setRegister(void* r, char reg, int val);
+int Registry_setRegister(void* r, char reg, unsigned long int val);
 int Registry_performOperation(void* r, char op, signed int* argv, int argc);
 int Registry_getRegisterValue(void* r, char reg, signed int* target);
 signed int* Registry_translateRegister(void* r, char regNum);
